Fix tuple query-builder friend, decouple sync headers from <coroutine>, rename .inl
Three related header-level fixes plus documentation:
 - The std::tuple specialization of SqlAllFieldsQueryBuilder declared its CRTP-base
   friend without the QueryOptions argument, so for any non-default DataMapperOptions
   the befriended type differed from the actual base and the base's RunFinisher lost
   access to the derived's private members (a latent hard compile error). Pass the
   missing argument, matching the primary template.
 - QueryBuilders.hpp unconditionally included Async/Backend.hpp, pulling <coroutine>
   and the full async runtime into every TU that includes DataMapper (including
   sync-only users), defeating the Async/Fwd.hpp decoupling. Move the heavy include
   to DataMapper.hpp (where RunFinisher's async branch needs it); QueryBuilders.hpp
   names no async type outside comments.
 - Rename Async/DataMapperAsync.inl to .hpp (no .inl extension) and update the
   include, the source list, and the install rule.
 - Strengthen the record/builder reference-capture lifetime contract in the header
   comment and docs/async.md: the operand must outlive the whole co_await, not just
   the call.
